fix: unittest organizational optimization
