+++ crates/bytesutil/Cargo.toml
@@ -6,5 +6,9 @@ edition = {workspace = true}
 [lints]
 workspace = true
 
+[lib]
+# no doctests currently implemented in this package; skip, then
+doctest = false
+
 [dependencies]
 bytes = { workspace = true }
