+++ crates/cli/Cargo.toml
@@ -10,6 +10,9 @@ workspace = true
 name = "glaredb"
 path = "src/bin/main.rs"
 
+[lib]
+doctest = false
+
 [dependencies]
 anyhow = { workspace = true }
 clap = { workspace = true }
@@ -50,14 +53,7 @@ tokio-postgres = "0.7.8"
 uuid = { version = "1.8.0", features = ["v4", "fast-rng", "macro-diagnostics"] }
 glob = "0.3.1"
 
-# Prevent dynamic linking of lzma, which comes from datafusion
-lzma-sys = { version = "*", features = ["static"] }
-
-[dev-dependencies]
-predicates = "3.1.0"
-assert_cmd = "2.0.14"
-tokio-postgres = "0.7.8"
-tempfile = { workspace = true }
+lzma-sys = { version = "*", features = ["static"] } # Prevent dynamic linking of lzma, which comes from datafusion
 
 [build-dependencies]
 built = { version = "0.7.2", features = ["git2"] }
