+++ crates/distexec/Cargo.toml
@@ -3,7 +3,10 @@ name = "distexec"
 version = { workspace = true }
 edition = { workspace = true }
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+# no tests currently implemented in this package; skip, then
+doctest = false
+test = false 
 
 [dependencies]
 thiserror.workspace = true
