+++ crates/glaredb/Cargo.toml
@@ -6,6 +6,11 @@ edition.workspace = true
 [lints]
 workspace = true
 
+[lib]
+# no tests currently implemented in this package; skip, then
+doctest = false
+test = false
+
 [dependencies]
 sqlexec = { path = "../sqlexec" }
 url = { workspace = true }
