+++ crates/logutil/Cargo.toml
@@ -6,6 +6,11 @@ edition = {workspace = true}
 [lints]
 workspace = true
 
+[lib]
+# no tests currently implemented in this package; skip, then
+doctest = false
+test = false
+
 [dependencies]
 tracing = { workspace = true }
 tracing-subscriber = {version = "0.3", features = ["std", "fmt", "json", "env-filter"] }
