+++ crates/parser/Cargo.toml
@@ -3,13 +3,14 @@ name = "parser"
 version.workspace = true
 edition.workspace = true
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+# no doctests currently implemented in this package; skip, then
+doctest = false
+
+[lints]
+workspace = true
 
 [dependencies]
 prql-compiler = "0.11.3"
 sqlparser = { version = "0.45.0", features = ["visitor"] }
-thiserror.workspace = true
-
-
-[lints]
-workspace = true
+thiserror = { workspace = true }
