+++ crates/protogen/Cargo.toml
@@ -6,6 +6,10 @@ edition.workspace = true
 [lints]
 workspace = true
 
+[lib]
+# no doctests currently implemented in this package; skip, then
+doctest = false
+
 [dependencies]
 datafusion = { workspace = true, features = ["serde"] }
 datafusion-proto = { workspace = true }
