+++ crates/sqlbuiltins/Cargo.toml
@@ -6,6 +6,10 @@ edition = { workspace = true }
 [lints]
 workspace = true
 
+[lib]
+# no doctests currently implemented in this package; skip, then
+doctest = false
+
 [dependencies]
 ioutil = { path = "../ioutil" }
 logutil = { path = "../logutil" }
