+++ crates/terminal_util/Cargo.toml
@@ -6,5 +6,10 @@ edition.workspace = true
 [lints]
 workspace = true
 
+[lib]
+# no tests currently implemented in this package; skip, then
+doctest = false
+test = false
+
 [dependencies]
 crossterm = "0.27.0"
