+++ rstests/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "rstests"
+version = { workspace = true }
+edition = { workspace = true }
+
+[lib]
+doctest = false
+
+[dev-dependencies]
+predicates = "3.1.0"
+assert_cmd = "2.0.14"
+tokio-postgres = "0.7.8"
+tempfile = { workspace = true }
