+++ rstests/tests/logging_test.rs
