+++ rstests/tests/output_mode_test.rs
